requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
openai==1.3.0
httpx==0.27.2
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0
reportlab==4.0.4
python-pptx==0.6.21
Pillow==10.0.0
requests==2.28.1
html2text==2025.4.15
psycopg2-binary==2.9.10
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface_hub==0.12.1
transformers==4.26.1
torch==2.0.1
torchvision==0.15.2
fpdf2==2.8.3
gunicorn==21.2.0 